- ⛽️🏌️‍♂️

+++ contracts/extensions/Deployer/Deployer_v001.sol
@@ -37,8 +37,8 @@ contract Deployer_v001 is OwnableUpgradeable, UUPSUpgradeable {
     uint256 _maxSupply,
     uint256 _unitPrice,
     uint256 _mintAllowance,
-    uint128 _mintPeriodStart,
-    uint128 _mintPeriodEnd
+    uint256 _mintPeriodStart,
+    uint256 _mintPeriodEnd
   ) external returns (address) {
     address t = NFTokenFactory.createNFToken(
       _owner,
